refactor(conflicts): extract the suppression predicates into the files their tests already name
Motion only; no predicate body changes, and every explanatory comment moves
with its code.

internal/conflicts contained branch_filters_test.go, disjoint_resource_test.go,
disjoint_work_item_test.go, operational_progression_test.go,
temporal_filter_test.go and cross_agent_precedent_test.go — six test files with
no source twin, because every one of those implementations lived inside
scorer.go. The package's own layout already said these are separate concerns;
only the source disagreed. The package carves logic this way everywhere else
(severity.go, claims.go, cost.go, crossencoder.go, ticket_extract.go,
recommendation.go, autoresolve.go, bindings.go, goldset.go).

Adding a suppression filter is the highest-value first contribution to this
detector. It now means editing a 70-to-170-line file named after the test the
contributor copied, instead of the package's hottest file at line 2123.

Nine files: branch_filters.go, disjoint_work_item.go, disjoint_resource.go,
operational_progression.go, temporal_filter.go, cross_agent_precedent.go,
workflow_filters.go, filter_vocabulary.go (the supersession, data-loss and
decision-type tables) and supersedes_suggestions.go (validatorSideDecision
through insertSupersedesSuggestion). scorer.go drops from 2,667 to 1,602 lines.

scoreForDecision is NOT split, deliberately. Its ten-gate ladder's ORDER is the
semantics; candidateScore is declared inside the function with a dozen fields
and closed over by both passes; the false-positive-rate cache is a closure over
ctx and orgID; suppressionSamples accumulates across the whole loop and flushes
once at the end. Scattering an ordered argument across three files, with
nothing recording the order, would destroy the only document explaining why a
given conflict was or was not flagged.

Two hazards, both handled:

  - scorer.go carries //go:build !lite on line 1. Every one of the nine new
    files repeats it verbatim; without that, ./cmd/akashi-local stops building.
    Verified with go build -tags lite -o /dev/null ./cmd/akashi-local. This
    commit does NOT make the predicates lite-visible by dropping the tag —
    that would change what the lite binary compiles and is a separate argument.

  - The provenance comments are the most valuable thing in this file: twelve
    drop sites, each naming the false-positive class and the issue that
    motivated it (#517, #692, #705, #709, #710, #717, the 2026-06 pgstream
    over-clustering audit, the 2026-06-23 disjoint-work-item class). Losing or
    shortening any of them would make this a regression rather than an
    improvement. Verified mechanically: every comment line present in the old
    scorer.go is still present across the new file set — the set difference is
    empty — and all 13 issue-referencing comments survive.

Verified as pure motion rather than asserted. The blocks were cut by a parser
that attaches each declaration's comment block and walks to its column-0
closing brace, and each destination file was re-parsed afterwards and
byte-compared against what was extracted: all nine identical. No test file was
touched. The package's 375 tests pass under -race, including
TestStructuralSuppressionRules_CallSiteParity, which is the guard that exists
precisely to catch a rule losing its sampling call site during an edit like
this one.
