+++ internal/conflicts/scorer.go
@@ -23,7 +23,6 @@ import (
 	"go.opentelemetry.io/otel/metric"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/ashita-ai/akashi/internal/compact"
 	"github.com/ashita-ai/akashi/internal/model"
 	"github.com/ashita-ai/akashi/internal/search"
 	"github.com/ashita-ai/akashi/internal/storage"
@@ -1565,1077 +1564,11 @@ func (s *Scorer) ClearAllConflicts(ctx context.Context) (int, error) {
 	return n, nil
 }
 
-// isComplementaryWorkflowPair returns true if the decision pair matches a
-// structural pattern where two decisions are about the same topic but are
-// complementary rather than contradictory. These patterns are invisible to
-// embedding-based scoring because topic similarity is high and outcome text
-// diverges (review language vs implementation language).
-//
-// Three heuristics, any of which is sufficient:
-//
-//  1. Temporal workflow: one decision is a review/assessment/investigation type
-//     and the other is an implementation/fix type, with the implementation
-//     recorded after the review. This covers review→fix and assessment→implementation.
-//
-//  2. Same-agent refinement: both decisions are from the same agent and the
-//     newer one's outcome contains keywords indicating it built on the older
-//     one ("implemented", "fixed", "resolved", "completed", "addressed").
-//
-//  3. Precedent chain: one decision cites the other via precedent_ref,
-//     meaning the agent explicitly linked them. However, precedent links can
-//     represent supersession (reversal) rather than refinement. When the later
-//     decision's outcome contains supersession keywords, the pair is passed
-//     through to LLM validation instead of being suppressed. See issue #452.
-func isComplementaryWorkflowPair(d, cand model.Decision) bool {
-	// Heuristic 3: Precedent chain. If either decision cites the other,
-	// they are linked by design. But linked decisions can still conflict
-	// (supersession). Check the later decision's outcome for reversal signals;
-	// if found, let the LLM decide instead of suppressing.
-	if isPrecedentLinked(d, cand) {
-		later := cand
-		if cand.ValidFrom.Before(d.ValidFrom) {
-			later = d
-		}
-		if containsSupersessionKeyword(later.Outcome) {
-			return false // probable supersession — let LLM validate
-		}
-		return true // probable refinement — suppress
-	}
-
-	// Determine temporal order: earlier and later decision.
-	earlier, later := d, cand
-	if cand.ValidFrom.Before(d.ValidFrom) {
-		earlier, later = cand, d
-	}
-
-	// Heuristic 1: Temporal workflow — review/assessment type followed by
-	// implementation/fix type.
-	if isDirectionalWorkflowPair(earlier.DecisionType, later.DecisionType) {
-		return true
-	}
-
-	// Heuristic 2: Same-agent refinement with outcome keywords.
-	if d.AgentID == cand.AgentID {
-		lowerOutcome := strings.ToLower(later.Outcome)
-		for _, kw := range refinementKeywords {
-			if strings.Contains(lowerOutcome, kw) {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-// isCoordinatedChange returns true when two decisions share provenance metadata
-// that proves they are part of the same coordinated change — same commit, same
-// PR, or same branch within a short temporal window. This is a binary signal
-// (no thresholds) that is semantically correct: decisions from the same PR are
-// coordinated by definition, even when they implement different layers (model,
-// handler, storage, UI, docs) whose outcome text diverges.
-//
-// Checked fields (via nestedContextString, which searches client.*, server.*,
-// and flat namespaces):
-//   - commit_sha: exact match → always coordinated (globally unique)
-//   - pr_number: exact match + same project → always coordinated
-//   - branch: exact match + same project + decisions within 24h → likely coordinated
-//
-// PR numbers and branch names are scoped to a repository, not globally unique.
-// Two decisions in the same org but different repos sharing pr_number "42" are
-// not coordinated. commit_sha is exempt because SHA-256 hashes are globally unique.
-//
-// Branch alone is a weaker signal (branches are reused across PRs), so it
-// requires temporal proximity as a secondary qualifier.
-func isCoordinatedChange(d, cand model.Decision) bool {
-	// Same commit SHA → definitively coordinated.
-	// Commit SHAs are globally unique, no project scoping needed.
-	commitA := nestedContextString(d.AgentContext, "commit_sha")
-	commitB := nestedContextString(cand.AgentContext, "commit_sha")
-	if commitA != "" && commitA == commitB {
-		return true
-	}
-
-	// For pr_number and branch checks, require same project. PR numbers and
-	// branch names are repo-scoped, not globally unique — pr_number "42" in
-	// repo-alpha is unrelated to pr_number "42" in repo-beta. When both
-	// projects are nil/empty we can't distinguish repos, so we require at
-	// least one non-empty project to avoid false suppression in multi-repo orgs.
-	projA := derefString(d.Project)
-	projB := derefString(cand.Project)
-	sameProject := projA == projB && (projA != "" || projB != "")
-
-	// Same PR number + same project → definitively coordinated.
-	prA := nestedContextString(d.AgentContext, "pr_number")
-	prB := nestedContextString(cand.AgentContext, "pr_number")
-	if prA != "" && prA == prB && sameProject {
-		return true
-	}
-
-	// Same branch + same project + temporal proximity → likely coordinated.
-	// 24h window accommodates multi-session work on a single branch
-	// while excluding branch reuse across separate PRs (which typically
-	// spans days/weeks with a merge in between).
-	branchA := nestedContextString(d.AgentContext, "branch")
-	branchB := nestedContextString(cand.AgentContext, "branch")
-	if branchA != "" && branchA == branchB && sameProject {
-		const coordinationWindow = 24 * time.Hour
-		timeDelta := d.ValidFrom.Sub(cand.ValidFrom).Abs()
-		if timeDelta <= coordinationWindow {
-			return true
-		}
-	}
-
-	return false
-}
-
-// mechanicalKeywords are outcome substrings that indicate a mechanical/routine
-// operation (migration renumbering, rebase conflict resolution, version bumps)
-// rather than a genuine design or architecture decision.
-var mechanicalKeywords = []string{
-	"renumber",
-	"renumbering",
-	"renumbered",
-	"rebase",
-	"rebasing",
-	"rebased",
-	"merge conflict",
-	"merge conflicts",
-	"merged origin/main",
-	"merged main",
-	"version bump",
-	"bumped version",
-}
-
-// isMechanicalOperation returns true when the outcome text describes a routine
-// mechanical operation like migration renumbering or rebase conflict resolution.
-// These operations produce semantically similar outcomes across branches but
-// represent parallel correct work, not disagreement.
-func isMechanicalOperation(outcome string) bool {
-	lower := strings.ToLower(outcome)
-	for _, kw := range mechanicalKeywords {
-		if strings.Contains(lower, kw) {
-			return true
-		}
-	}
-	return false
-}
-
-// isCrossBranchMechanical returns true when two decisions are on different git
-// branches and both describe mechanical operations (e.g. migration renumbering
-// during a rebase). These are parallel correct work whose resolution is
-// determined by merge order, not a genuine conflict.
-//
-// Requires both decisions to have branch metadata in agent_context and for the
-// branches to differ. Same-branch or missing-branch pairs are not affected.
-func isCrossBranchMechanical(d, cand model.Decision) bool {
-	branchA := nestedContextString(d.AgentContext, "git_branch")
-	branchB := nestedContextString(cand.AgentContext, "git_branch")
-
-	if branchA == "" || branchB == "" || branchA == branchB {
-		return false
-	}
-
-	return isMechanicalOperation(d.Outcome) && isMechanicalOperation(cand.Outcome)
-}
-
-// isSameBranchSelfCorrection returns true when the same agent made two
-// sequential decisions on the same branch — a self-correction pattern, not a
-// contradiction. The later decision supersedes the earlier one as part of
-// iterative work on the same branch.
-//
-// Requires both decisions to have branch metadata, be from the same agent,
-// and be on the same branch. The temporal order doesn't matter — the fact
-// that the same agent revised their own decision on the same branch is
-// sufficient to classify this as a self-correction.
-func isSameBranchSelfCorrection(d, cand model.Decision) bool {
-	if d.AgentID != cand.AgentID {
-		return false
-	}
-
-	branchA := nestedContextString(d.AgentContext, "git_branch")
-	branchB := nestedContextString(cand.AgentContext, "git_branch")
-
-	return branchA != "" && branchB != "" && branchA == branchB
-}
-
-// validatorSideDecision maps the side token the validator named on its REPLACES
-// line back onto the scored pair.
-//
-// The mapping is fixed by the ValidateInput literal in scoreForDecision:
-// OutcomeA / TypeA / AgentA / CreatedA / FullOutcomeA / BranchA / TicketRefsA
-// are ALL built from d, and every *B field from cand. So "A" is d and "B" is
-// cand. Any change to that literal must change this function with it; there is
-// no other place the correspondence is recorded.
-func validatorSideDecision(side string, d, cand model.Decision) (superseding, superseded model.Decision, ok bool) {
-	switch side {
-	case "A":
-		return d, cand, true
-	case "B":
-		return cand, d, true
-	default:
-		return model.Decision{}, model.Decision{}, false
-	}
-}
-
-// recordSupersedesSuggestion writes the detector-inferred supersedes link for a
-// suppressed same-agent same-ticket pair so akashi_check can surface it on the
-// agent's next call (PR-2 of #708, see issue #710).
-//
-// This path runs in the pre-LLM structural-filter block: there is no judge
-// verdict, so recorded order is the ONLY direction signal available and the
-// reason string says so explicitly. Do not borrow the LLM path's language here
-// — the attribution has to stay honest about what produced it.
-func (s *Scorer) recordSupersedesSuggestion(ctx context.Context, d, cand model.Decision, topicSim float64, ticket string) {
-	superseding, superseded := d, cand
-	if cand.ValidFrom.After(d.ValidFrom) {
-		superseding, superseded = cand, d
-	}
-	s.insertSupersedesSuggestion(ctx, superseding, superseded, topicSim,
-		supersedesSourceSameTicket,
-		fmt.Sprintf("same agent %q, same ticket %q; direction from recorded order (no validator verdict on this path)",
-			d.AgentID, ticket))
-}
-
-// recordSupersedesSuggestionFromValidator records the supersedes link the LLM
-// validator inferred when it classified a pair as a supersession. A supersession
-// is a lifecycle event — one decision explicitly replaces another — not a
-// standing disagreement between live positions, so it is surfaced as a supersedes
-// suggestion (which drives the supersedes_id link on the agent's next
-// akashi_check, issue #710) instead of being opened as a conflict a human must
-// triage.
-//
-// Direction comes from the judge's REPLACES line, never from ValidFrom. The
-// prompt refuses recency as evidence ("every candidate pair has a time order,
-// so ordering alone means nothing") and re-deriving direction from the clock
-// here reintroduced exactly that, breaking on backdated traces, late-filed
-// decisions, and reverts. When the judge's direction and recorded order
-// disagree, that is signal about the trace, not an error to resolve in favour
-// of the clock: the row is written with the judge's direction under a distinct
-// suggested_by so the two populations stay separable.
-func (s *Scorer) recordSupersedesSuggestionFromValidator(ctx context.Context, d, cand model.Decision, topicSim float64, result ValidationResult) {
-	superseding, superseded, ok := validatorSideDecision(result.SupersedingSide, d, cand)
-	if !ok {
-		// ParseValidatorResponse guarantees a supersession verdict carries a
-		// side token — it downgrades one that does not. So an empty or
-		// unrecognised side here means the ValidationResult did not come from
-		// the parser: an in-process validator, or a test double constructing
-		// the struct directly.
-		//
-		// Direction IS the content of this record, so refusing to write it is
-		// the only honest option; falling back to ValidFrom is the clock
-		// inference this function exists to remove. Logged at error rather than
-		// warn because it is a contract violation, not a transient write
-		// failure — but it still returns rather than propagating, because this
-		// is a fire-and-forget path inside the scoring loop and the documented
-		// contract is that suggestion writes never block scoring. Loud with no
-		// side effect beats silent with a guessed direction.
-		s.metrics.supersessionSideMissing.Add(ctx, 1)
-		s.logger.Error("conflict scorer: supersession verdict carried no superseding side, no supersedes suggestion written",
-			"decision_a", d.ID, "decision_b", cand.ID,
-			"side", result.SupersedingSide,
-			"validator", s.validatorLabel)
-		return
-	}
-
-	// The judge quoted replacement language when it could; its one-sentence
-	// explanation is the fallback. Either way the reason carries the judge's
-	// own finding instead of a template rendered from an inference we just made.
-	detail := result.ReplacementEvidence
-	if detail == "" {
-		detail = result.Explanation
-	}
-	detail = compact.Truncate(strings.TrimSpace(detail), supersedesReasonMaxDetail)
-
-	// Decision IDs are included because agent IDs do not disambiguate the sides
-	// of a same-agent pair, which is the common case for this detector.
-	reason := fmt.Sprintf("LLM validator: %q's decision %s replaces %q's decision %s — %s",
-		superseding.AgentID, superseding.ID.String()[:8],
-		superseded.AgentID, superseded.ID.String()[:8], detail)
-
-	suggestedBy := supersedesSourceLLM
-	// Disagreement is strict: equal timestamps mean the clock has no opinion,
-	// not that it dissents.
-	if superseded.ValidFrom.After(superseding.ValidFrom) {
-		suggestedBy = supersedesSourceLLMBackdated
-		reason = "judge direction disagrees with recorded order (the retired decision has the later valid_from): " + reason
-		s.metrics.supersessionDirectionDisagreement.Add(ctx, 1)
-		s.logger.Warn("conflict scorer: supersedes direction from validator disagrees with recorded order",
-			"superseding_id", superseding.ID, "superseded_id", superseded.ID,
-			"superseding_valid_from", superseding.ValidFrom,
-			"superseded_valid_from", superseded.ValidFrom,
-			"side", result.SupersedingSide,
-			"validator", s.validatorLabel)
-	}
-
-	s.insertSupersedesSuggestion(ctx, superseding, superseded, topicSim, suggestedBy, reason)
-}
-
-// insertSupersedesSuggestion is the shared core behind the supersedes-suggestion
-// callers. Direction is the CALLER's decision — this function does not infer it
-// — because the two callers derive it from different evidence (the judge's
-// REPLACES line vs recorded order on a path with no judge). It clamps topicSim
-// into the [0,1] confidence contract (cosine similarity is mathematically in
-// [-1,1] but the API bounds confidence to [0,1]) and writes fire-and-forget: a
-// write failure is logged at warn and never blocks scoring, because the pair has
-// already been filtered or redirected away from the open queue by the time this
-// runs.
-func (s *Scorer) insertSupersedesSuggestion(ctx context.Context, superseding, superseded model.Decision, topicSim float64, suggestedBy, reason string) {
-	switch {
-	case topicSim < 0:
-		topicSim = 0
-	case topicSim > 1:
-		topicSim = 1
-	}
-	conf := float32(topicSim)
-	inverseExists, err := s.db.InsertSupersedesSuggestion(ctx, storage.SupersedesSuggestionInsert{
-		OrgID:         superseding.OrgID,
-		SupersedingID: superseding.ID,
-		SupersededID:  superseded.ID,
-		SuggestedBy:   suggestedBy,
-		Confidence:    &conf,
-		Reason:        reason,
-	})
-	if err != nil {
-		s.logger.Warn("conflict scorer: insert supersedes suggestion failed",
-			"superseding_id", superseding.ID, "superseded_id", superseded.ID,
-			"suggested_by", suggestedBy, "error", err)
-		return
-	}
-	if inverseExists {
-		// The opposite direction is already recorded for this pair. Since the
-		// judge now decides direction and is sampled, this means two passes over
-		// the same pair disagreed about which decision retired which. Nothing was
-		// written. Surface it at WARN rather than resolving it silently — a
-		// direction picked by whichever pass ran last is exactly the guessed link
-		// this change exists to stop recording.
-		s.metrics.supersedesInverseSuppressed.Add(ctx, 1)
-		s.logger.Warn("conflict scorer: supersedes suggestion contradicts an existing inverse link, not recorded",
-			"superseding_id", superseding.ID, "superseded_id", superseded.ID,
-			"suggested_by", suggestedBy)
-	}
-}
-
-// isSameAgentSameTicketRefinement returns true when the same agent traces
-// what looks like a clean refinement of their own prior decision on the same
-// ticket without setting supersedes_id. The newer decision likely intends
-// to supersede the earlier one but the agent forgot to set the link.
-//
-// Sibling of isSameBranchSelfCorrection: that filter requires identical
-// git_branch metadata. This one is broader — refinements can span branches
-// (e.g. layer-2 PR followed by layer-3 PR on the same ticket) so the join key
-// is an extracted ticket reference (see extractTicketRef in ticket_extract.go).
-//
-// Excluded:
-//   - precedent-linked pairs (the agent already linked them — let LLM judge)
-//   - outcomes containing supersession keywords (probable explicit reversal,
-//     not a forgotten link — let LLM judge)
-//   - different agents
-//   - missing or differing ticket references
-//
-// Tracks issue #709 (PR-1 of #708). PR-2 (#710) will surface a supersedes_id
-// suggestion to the agent via akashi_check.
-func isSameAgentSameTicketRefinement(d, cand model.Decision) bool {
-	if d.AgentID != cand.AgentID {
-		return false
-	}
-	if isPrecedentLinked(d, cand) {
-		return false
-	}
-	later := d
-	if cand.ValidFrom.After(d.ValidFrom) {
-		later = cand
-	}
-	if containsSupersessionKeyword(later.Outcome) {
-		return false
-	}
-	refA := extractTicketRef(d)
-	if refA == "" {
-		return false
-	}
-	return refA == extractTicketRef(cand)
-}
-
-// isCrossAgentPrecedentRefinement returns true when two different agents
-// trace decisions on the same ticket, one explicitly cites the other via
-// precedent_ref, and the later outcome contains no supersession keywords.
-//
-// Cross-agent analogue of isSameAgentSameTicketRefinement: the same-agent
-// filter targets forgotten-supersedes within one agent's work; this filter
-// targets the case where agent B explicitly declares "I am building on A"
-// (by setting precedent_ref to A on the same ticket) yet the LLM validator
-// still classifies the pair as contradiction because the two outcomes
-// mention competing implementation details.
-//
-// Conservative by construction. Required signals:
-//   - different agents
-//   - precedent_ref explicit between the two decisions (either direction)
-//   - identical extracted ticket reference on both sides
-//   - same project (defense against pathological cross-project precedent_refs)
-//   - later outcome free of supersession keywords (so declared reversals
-//     and partial supersessions still reach the validator)
-//
-// Does NOT catch (deliberately):
-//   - same-ticket cross-agent pairs without precedent_ref — those may be
-//     genuine disagreements (e.g. ARD-1168 broad vs narrow shared-env
-//     fallback, which resolved with codex winning)
-//   - pairs sharing only a common-ancestor precedent_ref but not citing
-//     each other — shared ancestry is too weak a signal to suppress, in
-//     line with the 8f5124c8 reversal of bare same-branch suppression
-//
-// See sibling isSameAgentSameTicketRefinement for the same-agent variant.
-func isCrossAgentPrecedentRefinement(d, cand model.Decision) bool {
-	if d.AgentID == cand.AgentID {
-		return false
-	}
-	if !isPrecedentLinked(d, cand) {
-		return false
-	}
-	if derefString(d.Project) != derefString(cand.Project) {
-		return false
-	}
-	later := d
-	if cand.ValidFrom.After(d.ValidFrom) {
-		later = cand
-	}
-	if containsSupersessionKeyword(later.Outcome) {
-		return false
-	}
-	refA := extractTicketRef(d)
-	if refA == "" {
-		return false
-	}
-	return refA == extractTicketRef(cand)
-}
-
 // temporalReassessmentWindow is defined in shared.go — it is referenced by the
 // lite-compiled validator.go and so must live in an untagged file.
 
-// isTemporalReassessment returns true when two decisions are both
-// review/assessment types on the same project, separated by at least
-// temporalReassessmentWindow, with neither citing the other and not in
-// the same session. This is the re-measurement pattern: the newer
-// assessment updates a prior measurement of the same system and naturally
-// has different numbers without contradicting the earlier conclusions.
-//
-// Pairs caught by this filter would otherwise embed close together (same
-// domain) and produce a CONTRADICTION verdict from the LLM whenever quoted
-// metrics differ — see issue #705. Linked or same-session re-assessments
-// are deliberately excluded so the LLM can still classify intentional
-// supersessions.
-func isTemporalReassessment(d, cand model.Decision) bool {
-	if !reviewTypes[strings.ToLower(d.DecisionType)] {
-		return false
-	}
-	if !reviewTypes[strings.ToLower(cand.DecisionType)] {
-		return false
-	}
-	if derefString(d.Project) != derefString(cand.Project) {
-		return false
-	}
-	if isPrecedentLinked(d, cand) {
-		return false
-	}
-	if d.SessionID != nil && cand.SessionID != nil && *d.SessionID == *cand.SessionID {
-		return false
-	}
-	delta := d.ValidFrom.Sub(cand.ValidFrom)
-	if delta < 0 {
-		delta = -delta
-	}
-	return delta >= temporalReassessmentWindow
-}
-
-// operationalProgressionWindow is the minimum time delta between two
-// operational decisions on the same project below which the pair is never
-// treated as sequential progression — near-simultaneous competing directives
-// are the genuine operational-clash shape and must stay detectable. Beyond the
-// window, a large gap is a *weak* signal that the system state has moved on and
-// the earlier action no longer describes what the later one is changing.
-//
-// The signal is only weak, and deliberately not relied on alone: a single
-// resource can stay in one incident longer than the window (the live Acme
-// connector_redacted01 incident ran 2026-05-26 → 06-18), in which case directives
-// spanning >7 days are still about the same live system, not unrelated steps.
-// Time-apart cannot distinguish that case from genuine progression, so
-// isOperationalStateProgression layers other guards on top: the data-loss guard
-// exempts high-stakes pairs outright regardless of gap, and the
-// precedent/session/supersession guards exempt declared lineage. Matches
-// temporalReassessmentWindow (7 days = the SessionStart "recent" window): the
-// same "state moved on" rationale applied to operational state instead of
-// measured metrics.
-const operationalProgressionWindow = 7 * 24 * time.Hour
-
-// isOperationalStateProgression returns true when two decisions are both
-// operational-execution types on the same project, separated by at least
-// operationalProgressionWindow, with neither framed as a reversal of a prior
-// approach and no explicit precedent link between them.
-//
-// Motivation: the dominant live false-positive class (2026-06 open-conflict
-// audit) was cross-ticket operational steps on one hot subsystem
-// (pgstream / Debezium / Acme CDC) clustered by shared vocabulary —
-// e.g. "rolled kafka2pg back to its pre-rollout digest" eight days after
-// "verified the connector ONLINE". Embeddings place them close (topic_sim
-// 0.70-0.85) and the LLM validator returns CONTRADICTION, but a rollback or
-// pause executed a week after a prior deploy is the next step in an incident
-// timeline, not a disagreement with it.
-//
-// Operational sibling of isTemporalReassessment: that filter covers
-// review-type re-measurements whose numbers drift over time; this one covers
-// operational state that mutates over time. Both rest on the same principle —
-// an action taken long ago does not contradict a later one just because they
-// touch the same system.
-//
-// Deliberately narrow, to keep genuine conflicts detectable:
-//   - both decisions must be operationalTypes. architecture / trade_off /
-//     design forks (e.g. the real Debezium-vs-pgstream disagreement) are
-//     never eligible, so they still reach the validator. This is why
-//     trade_off and bug_fix are excluded even though some operational FPs
-//     carry those types — widening would risk suppressing real design forks.
-//   - same non-empty project (untagged-vs-untagged pairs are not suppressed)
-//   - >= operationalProgressionWindow apart, so near-simultaneous competing
-//     directives (the genuine operational-clash shape) stay detectable
-//   - neither outcome contains a supersession keyword — an action that
-//     explicitly reverts/abandons a prior approach is a declared reversal and
-//     reaches the LLM (checked on both sides: the safe, under-suppressing
-//     direction)
-//   - neither side's task or outcome reports a data-loss / corruption /
-//     quarantine event — a data-safety event is categorically high-stakes and
-//     must reach the validator regardless of the time gap (checked on both
-//     sides, both text fields). This is the guard that keeps same-resource
-//     incident disagreements like the live Acme DATALOSS pause from
-//     being silently dropped pre-LLM. See decisionContainsDataLossKeyword.
-//   - not precedent-linked and not the same session — explicit links and
-//     intra-session pairs are left to the LLM, mirroring isTemporalReassessment
-//
-// Known limitation: suppression keys on (type, project, gap, keywords) with no
-// resource-identity signal, so a same-resource contradiction that mentions no
-// data-safety event can still be suppressed once it is past the window. The
-// data-loss guard closes the high-stakes slice of that gap — the only slice
-// with verified instances in the trail. The complementary different-resource
-// slice (and operational pairs inside the same incident window, which this
-// filter's gap requirement skips) is now handled pre-LLM by isDisjointResource,
-// which keys on the structured connector_/org_ tokens this filter declined to
-// parse — those tokens are regular identifiers, not the fragile free-text
-// customer names that motivated the original deferral.
-//
-// Scope: lives only in the cloud scorer (this file is !lite). The lite scorer
-// has none of the structural suppression family; porting it is out of scope.
-func isOperationalStateProgression(d, cand model.Decision) bool {
-	if !operationalTypes[strings.ToLower(d.DecisionType)] {
-		return false
-	}
-	if !operationalTypes[strings.ToLower(cand.DecisionType)] {
-		return false
-	}
-	projA, projB := derefString(d.Project), derefString(cand.Project)
-	if projA != projB || (projA == "" && projB == "") {
-		return false
-	}
-	if isPrecedentLinked(d, cand) {
-		return false
-	}
-	if d.SessionID != nil && cand.SessionID != nil && *d.SessionID == *cand.SessionID {
-		return false
-	}
-	if containsSupersessionKeyword(d.Outcome) || containsSupersessionKeyword(cand.Outcome) {
-		return false
-	}
-	// Data-safety guard: never auto-suppress a pair where either side reports a
-	// data-loss / corruption / quarantine event. Such a decision is not a routine
-	// sequential step, and a same-system pair that includes one is precisely the
-	// disagreement that must reach the validator and the conflict queue rather
-	// than be dropped pre-LLM with only a Debug log. Both sides checked, across
-	// task and outcome (under-suppressing direction), mirroring the supersession
-	// guard above. (The supersession guard stays outcome-only on purpose: a
-	// reversal verb in a task title — "migrate connector from kafka to pg" — is
-	// routine phrasing, and scanning the task for it would re-admit the very FPs
-	// this filter kills. The data-safety guard has no such cost: firing it only
-	// ever routes a pair to the validator, never suppresses one.)
-	if decisionContainsDataLossKeyword(d) || decisionContainsDataLossKeyword(cand) {
-		return false
-	}
-	delta := d.ValidFrom.Sub(cand.ValidFrom)
-	if delta < 0 {
-		delta = -delta
-	}
-	return delta >= operationalProgressionWindow
-}
-
-// isDisjointWorkItem returns true when two work-item-scoped decisions (reviews,
-// plans) on the same project reference entirely different work items — different
-// PRs or different tickets — and therefore cannot be contradicting each other.
-// A review of PR #1539 and a review of PR #1543 examine different code; a plan
-// for ARD-1606 and a plan for ARD-1662 scope different tickets. They routinely
-// share dense subsystem vocabulary (cutover, pgstream, freeze, replica-identity,
-// shadow-apply) that places them close in embedding space, and the LLM validator
-// then manufactures a CONTRADICTION between them.
-//
-// This is the hard-rule escalation of the disjoint-ticket signal that
-// eef7eb2c shipped as a *validator prompt hint* ("DIFFERENT TICKETS ... classify
-// as UNRELATED or COMPLEMENTARY unless ..."). The live 2026-06-23 open-conflict
-// audit showed the hint is ignored once shared vocabulary is dense enough: 9 of
-// 19 open false positives were disjoint-work-item review/planning pairs the
-// prompt failed to deflect. Promoting it to a structural pre-LLM suppressor is
-// the membership-parity follow-through, and — unlike the ticket-only hint — this
-// one is PR-aware via extractWorkItemRefs, which is what closes the dominant
-// fan-out hole: the hub decision in that audit (a review of "PR #1539") carries
-// NO ARD-style ticket ref at all, so a ticket-only rule could never have matched
-// the nine pairs it spawned.
-//
-// Deliberately narrow, to keep genuine conflicts detectable:
-//   - both decisions must be workItemScopedTypes. architecture / trade_off /
-//     design forks are never eligible — a cross-cutting design disagreement
-//     (e.g. the real Debezium-vs-pgstream fork) can legitimately span work
-//     items and must reach the validator. Same exclusion rationale as
-//     isOperationalStateProgression.
-//   - same non-empty project (PR/ticket numbers are repo-scoped; an untagged
-//     pair, or a cross-repo pair sharing a PR number, is never suppressed —
-//     mirrors isCoordinatedChange's same-project requirement for pr_number)
-//   - both decisions must expose at least one extractable work-item ref, and
-//     the two ref sets must be fully disjoint. Any shared PR/ticket (including
-//     one outcome naming the other's work item, e.g. "closed #1543 in favor of
-//     #1545") makes the sets overlap and the pair reaches the validator.
-//   - not precedent-linked — an explicit lineage cite is left to the LLM
-//   - neither side's task or outcome reports a data-loss / corruption /
-//     quarantine event — a data-safety finding is categorically high-stakes and
-//     must never be dropped pre-LLM (both sides, both text fields — the same
-//     task+outcome extractWorkItemRefs mines for the work-item refs that make
-//     the pair eligible, so a keyword cannot hide in a field the ref scan reads
-//     but the guard does not). This is NOT a redundant copy of the operational
-//     filter's guard: cross-ticket data-safety contradictions are a real,
-//     resolved class in this trail (e.g. two investigations root-causing the
-//     same DATALOSS incident to different causes under different tickets), and
-//     disjoint work items do not make that pair safe to drop. Same guard, same
-//     reasoning as isOperationalStateProgression; see
-//     decisionContainsDataLossKeyword.
-//
-// Deliberately NOT guarded on supersession keywords, unlike the operational and
-// refinement siblings. Those siblings have no work-item identity signal, so a
-// "replaced X" / "superseded by Y" phrase is their only cue that a pair might be
-// a declared reversal that should reach the validator. This filter HAS that
-// signal: a genuine cross-work-item supersession names the other work item, so
-// the ref sets overlap and the pair already reaches the validator (then the
-// post-LLM supersession→suggestion path). Adding a keyword guard here would only
-// re-admit the false positives this filter exists to kill — review prose is full
-// of code-diff verbs ("replaced the boolean", "dropped the dead reader") that
-// describe a diff, not a decision reversal. A vague reversal that names no work
-// item is the sole edge this forgoes; it stays suppressed, which is correct (an
-// unnamed antecedent cannot be linked anyway). The difference from the siblings
-// is intentional, not drift.
-//
-// Failure mode is under-suppression: a missed PR/ticket extraction leaves the
-// ref sets looking smaller (or one side empty), which disables the filter and
-// sends the pair to the validator — never a wrongful suppression.
-//
-// Scope: lives only in the cloud scorer (this file is !lite), alongside the rest
-// of the structural suppression family.
-func isDisjointWorkItem(d, cand model.Decision) bool {
-	if !workItemScopedTypes[strings.ToLower(d.DecisionType)] {
-		return false
-	}
-	if !workItemScopedTypes[strings.ToLower(cand.DecisionType)] {
-		return false
-	}
-	projA, projB := derefString(d.Project), derefString(cand.Project)
-	if projA != projB || (projA == "" && projB == "") {
-		return false
-	}
-	if isPrecedentLinked(d, cand) {
-		return false
-	}
-	if decisionContainsDataLossKeyword(d) || decisionContainsDataLossKeyword(cand) {
-		return false
-	}
-	refsA := extractWorkItemRefs(d)
-	refsB := extractWorkItemRefs(cand)
-	if len(refsA) == 0 || len(refsB) == 0 {
-		return false
-	}
-	return !ticketRefsOverlap(refsA, refsB)
-}
-
-// isDisjointResource returns true when two operational / investigation decisions
-// on the same project act on entirely different infrastructure resources —
-// different connector_/org_ identifiers — and therefore cannot contradict each
-// other. A diagnosis of connector_57a42328 and a recovery of connector_9b38b47d
-// touch different data planes; they routinely share dense incident vocabulary
-// (restart_storm, kafka2pg, replication slot, wal_status, snapshot) that places
-// them close in embedding space, and the validator then manufactures a
-// CONTRADICTION between two unrelated incidents.
-//
-// This is the resource-identity sibling of isDisjointWorkItem. That filter keys
-// on PR/ticket refs, which live operational traces do not carry — they name a
-// connector_/org_ token instead — so it could never see this class. Together
-// they are the membership-parity pair: work-item identity for code reviews and
-// plans, resource identity for incident recovery. It is the dominant live FP
-// class (2026-06-24 cross-connector restart-storm over-clustering) and the
-// reason operational conflicts pile up despite the work-item filter.
-//
-// It also fills the resource-identity gap that isOperationalStateProgression
-// documents as deferred: that filter suppresses same-project operational pairs
-// only once they are >= operationalProgressionWindow apart, so different-resource
-// pairs inside the same incident window (hours apart) slip past it. This filter
-// has no time requirement — different resources cannot contradict regardless of
-// when the two decisions were traced.
-//
-// Deliberately narrow, with the same discipline as isDisjointWorkItem:
-//   - both decisions must be resourceScopedTypes. architecture / trade_off /
-//     design forks are never eligible — a cross-cutting design disagreement can
-//     legitimately span resources and must reach the validator.
-//   - same non-empty project (operational traces are all project "mono"; an
-//     untagged pair is never suppressed), mirroring the sibling filters.
-//   - both must expose at least one extractable resource ref, and the two ref
-//     sets must be provably disjoint by namespace (see
-//     resourceRefsProvablyDisjoint). Disjointness is only judged WITHIN a
-//     namespace: a different connector vs connector, or org vs org, is disjoint;
-//     a connector vs an org is NOT comparable — the connector may be owned by
-//     that org, and the scorer holds no connector→org mapping — so that pair
-//     reaches the validator. Any shared connector/org, or any namespace present
-//     on only one side, also sends the pair to the validator — a shared
-//     resource is exactly where a genuine operational clash lives.
-//   - not precedent-linked — an explicit lineage cite is left to the LLM.
-//
-// No data-safety guard here (removed 2026-07-23). The guard is load-bearing and
-// deliberately chosen in the TEMPORAL sibling isOperationalStateProgression
-// (decision efc42730): two decisions on the SAME writer a week apart can be one
-// incident disagreeing with itself — verified against the Acme
-// connector_redacted01 DATALOSS pause. That rationale cannot hold here, because a
-// connector_/org_ id is the PHYSICAL identity of a data plane: connector_57a42328
-// and connector_f924df29 are different incidents by construction, so a DATALOSS
-// finding on one can neither be the same incident as, nor contradict, a finding
-// on the other. The guard was copied here for "membership parity" with the
-// temporal filter, but it only ever governed provably-disjoint pairs (this
-// function returns true solely when resourceRefsProvablyDisjoint) and, measured
-// against the live corpus, was a false-positive driver: "quarantine"/"corrupt"
-// are ordinary pgstream design vocabulary, so the guard re-admitted the exact
-// connector-disjoint over-clustering this filter exists to kill. Same-writer
-// data-safety disagreements stay protected where the verified evidence put them,
-// in isOperationalStateProgression. (isDisjointWorkItem keeps its guard on
-// purpose: a ticket is an administrative label, not a physical identity, so two
-// disjoint tickets CAN describe one incident — a deliberate, documented
-// difference, not drift.)
-//
-// Failure mode is under-suppression: a decision that names its resource only by
-// customer name ("Trellis") and not by a connector_/org_ token yields no ref,
-// which disables the filter and sends the pair to the validator. So does a
-// cross-namespace pair (a connector on one side, an org on the other) and any
-// pair sharing a resource — never a wrongful suppression, because disjointness
-// is only ever declared between same-namespace identifiers that genuinely
-// differ.
-//
-// Scope: lives only in the cloud scorer (this file is !lite), alongside the rest
-// of the structural suppression family.
-func isDisjointResource(d, cand model.Decision) bool {
-	if !resourceScopedTypes[strings.ToLower(d.DecisionType)] {
-		return false
-	}
-	if !resourceScopedTypes[strings.ToLower(cand.DecisionType)] {
-		return false
-	}
-	projA, projB := derefString(d.Project), derefString(cand.Project)
-	if projA != projB || (projA == "" && projB == "") {
-		return false
-	}
-	if isPrecedentLinked(d, cand) {
-		return false
-	}
-	refsA := extractResourceRefs(d)
-	refsB := extractResourceRefs(cand)
-	if len(refsA) == 0 || len(refsB) == 0 {
-		return false
-	}
-	return resourceRefsProvablyDisjoint(refsA, refsB)
-}
-
-// resourceRefsProvablyDisjoint reports whether two non-empty canonical
-// resource-reference sets ("CONNECTOR-<hex>", "ORG-<hex>") describe provably
-// different infrastructure — so a pair naming only these resources cannot be
-// acting on the same data plane and is safe to suppress pre-LLM.
-//
-// Disjointness is only provable WITHIN a namespace. CONNECTOR-* names a single
-// pipeline; ORG-* names a whole tenant, which owns many connectors. Two
-// different connector ids are different pipelines, and two different org ids are
-// different tenants — both provably disjoint. But a connector id and an org id
-// are NOT comparable: the connector may be owned by that very org, and the
-// scorer holds no connector→org mapping to rule it out. The original check
-// flattened both namespaces into one set and ran a plain overlap test, so a
-// decision naming only connector_57a42328 and one naming only its owning
-// org_54b2e846 looked "disjoint" (no string in common) and were suppressed —
-// even though they may describe the same incident on the same data plane. That
-// is a wrongful pre-LLM suppression, the one failure mode this family exists to
-// avoid.
-//
-// The fix types the refs by namespace and declares disjointness only when the
-// comparison is apples-to-apples in every namespace:
-//   - both sides must populate exactly the same set of namespaces. A
-//     connector-only side and an org-only side — or a connector-only side and a
-//     connector+org side — are not comparable, because there is a ref on one
-//     side whose ownership relative to the other side is unknown, so the pair
-//     reaches the validator.
-//   - within every populated namespace the two id sets must be fully disjoint.
-//     Any shared connector or shared org makes the pair overlap and reach the
-//     validator.
-//
-// This is strictly more conservative than the flattened check: it can only send
-// MORE pairs to the validator, never fewer (the safe, under-suppressing
-// direction the whole family is tuned to). The dominant live FP class —
-// connector-vs-connector, both sides naming only connectors — is unaffected:
-// same namespace set {CONNECTOR}, disjoint ids, still suppressed.
-func resourceRefsProvablyDisjoint(refsA, refsB []string) bool {
-	byNsA := groupResourceRefsByNamespace(refsA)
-	byNsB := groupResourceRefsByNamespace(refsB)
-	if len(byNsA) != len(byNsB) {
-		return false // different namespaces populated → not comparable
-	}
-	for ns, idsA := range byNsA {
-		idsB, ok := byNsB[ns]
-		if !ok {
-			return false // a namespace present on one side only → not comparable
-		}
-		for id := range idsA {
-			if _, shared := idsB[id]; shared {
-				return false // a shared resource in this namespace → overlap
-			}
-		}
-	}
-	return true
-}
-
-// groupResourceRefsByNamespace partitions canonical resource refs into id sets
-// keyed by their namespace prefix ("CONNECTOR", "ORG"). extractResourceRefs
-// produces every ref as "<PREFIX>-<hex>" with exactly one hyphen, so the split
-// is unambiguous; a malformed ref carrying no hyphen is skipped rather than
-// mis-bucketed.
-func groupResourceRefsByNamespace(refs []string) map[string]map[string]struct{} {
-	byNs := make(map[string]map[string]struct{}, 2)
-	for _, ref := range refs {
-		ns, id, ok := strings.Cut(ref, "-")
-		if !ok {
-			continue
-		}
-		ids, exists := byNs[ns]
-		if !exists {
-			ids = make(map[string]struct{}, 1)
-			byNs[ns] = ids
-		}
-		ids[id] = struct{}{}
-	}
-	return byNs
-}
-
-// isPrecedentLinked returns true if either decision cites the other via
-// precedent_ref, meaning there is an explicit lineage link between them.
-func isPrecedentLinked(d, cand model.Decision) bool {
-	if cand.PrecedentRef != nil && *cand.PrecedentRef == d.ID {
-		return true
-	}
-	if d.PrecedentRef != nil && *d.PrecedentRef == cand.ID {
-		return true
-	}
-	return false
-}
-
-// supersessionKeywords are outcome substrings that indicate the decision
-// reverses or replaces a prior decision rather than refining it.
-//
-// The "walk-back" group (shelved, dropped, pivoting, …) was added after the
-// #717 FP audit observed agents stating their own reversal in plain English
-// — "Shelved the draft; pivoting to a different topic" — without the
-// existing same-agent same-ticket filter recognising the reversal because
-// the keyword list missed the verb.
-var supersessionKeywords = []string{
-	"switched",
-	"superseding",
-	"superseded",
-	"replaced",
-	"replacing",
-	"reversed",
-	"reversing",
-	"reverted",
-	"reverting",
-	"migrated from",
-	"migrating from",
-	"instead of",
-	"no longer",
-	"abandoned",
-	"abandoning",
-	// Walk-back / pivot vocabulary surfaced by issue #717 FP audit.
-	"shelved",
-	"shelving",
-	"dropped",
-	"dropping",
-	"pivoting",
-	"pivoted",
-	"walked back",
-	"walking back",
-	"rewrote",
-	"rewriting",
-	"rewritten",
-	"scrapped",
-	"scrapping",
-	"discarded",
-	"discarding",
-	"retracted",
-	"retracting",
-	"withdrew",
-	"withdrawing",
-	"tabled",
-	"parked",
-	"deprecated this",
-	"deprecating",
-}
-
-// containsSupersessionKeyword returns true if the outcome text contains any
-// keyword indicating the decision reverses or replaces a prior one.
-func containsSupersessionKeyword(outcome string) bool {
-	lower := strings.ToLower(outcome)
-	for _, kw := range supersessionKeywords {
-		if strings.Contains(lower, kw) {
-			return true
-		}
-	}
-	return false
-}
-
-// dataLossKeywords are outcome substrings that mark an operational decision as
-// describing a data-safety event (loss, corruption, or quarantined/skipped
-// writes) rather than a routine state change. isOperationalStateProgression
-// refuses to suppress any pair where either side carries one: a data-safety
-// event is categorically high-stakes, and silently dropping a same-system pair
-// that reports one — before the LLM validator or a human ever sees it — would
-// trade the system's core guarantee (a complete paper trail of disagreements)
-// for noise reduction. Surfaced by the 2026-06 Acme connector_redacted01
-// incident, where "paused … active target-side DATALOSS quarantine growth" sat
-// >7 days from "scaled … resume loses no data" on the same writer.
-//
-// Deliberately broad and matched on both sides (the under-suppressing
-// direction): even a reassuring "verified no data loss" disables suppression
-// for the pair, which is correct — if data safety is in question at all, the
-// pair belongs in front of the validator.
-var dataLossKeywords = []string{
-	"dataloss",
-	"data loss",
-	"data-loss",
-	"lost data",
-	"data was lost",
-	"quarantine", // pgstream parks unappliable rows in a quarantine table
-	"corrupt",    // corrupted / corruption
-}
-
-// containsDataLossKeyword returns true if the given text references a
-// data-safety event (loss, corruption, or quarantined writes).
-func containsDataLossKeyword(text string) bool {
-	lower := strings.ToLower(text)
-	for _, kw := range dataLossKeywords {
-		if strings.Contains(lower, kw) {
-			return true
-		}
-	}
-	return false
-}
-
-// decisionContainsDataLossKeyword reports whether a decision references a
-// data-safety event in any free-text field that a pre-LLM suppressor reads to
-// judge eligibility — the agent-supplied agent_context.task and the outcome.
-//
-// The pre-LLM suppressors mine BOTH fields for their identity signal:
-// extractResourceRefs and extractWorkItemRefs both scan task and outcome, so a
-// pair can become eligible for suppression on the strength of a ref that appears
-// only in the task. The data-safety guard must therefore scan exactly those same
-// sources — otherwise a "DATALOSS recovery connector_x" task whose outcome omits
-// the word is silently dropped pre-LLM while its connector ref still drives the
-// disjoint-resource match. Guarding only the outcome (the original behavior) left
-// that hole. This is the membership-parity fix for the guard: every suppressor
-// that mines the task for identity now guards the task for data safety, so the
-// guard and the eligibility signal never read different text.
-//
-// isOperationalStateProgression shares this guard too, though it keys on a time
-// gap rather than task-extracted refs: a data-safety event anywhere in a
-// decision's salient text is categorically high-stakes and must reach the
-// validator, and a single uniform guard is one fewer place for the data-loss
-// check to drift between the sibling filters.
-func decisionContainsDataLossKeyword(d model.Decision) bool {
-	return containsDataLossKeyword(nestedContextString(d.AgentContext, "task")) ||
-		containsDataLossKeyword(d.Outcome)
-}
-
 // reviewTypes is defined in shared.go — it is referenced by the lite-compiled
 // validator.go and so must live in an untagged file.
-
-// implementationTypes are decision types that represent fix/implementation work.
-var implementationTypes = map[string]bool{
-	"architecture":   true,
-	"bug_fix":        true,
-	"fix":            true,
-	"implementation": true,
-	"refactor":       true,
-}
-
-// operationalTypes are decision types that represent execution actions against
-// running infrastructure (deploys, rollbacks, scaling, restarts, image
-// promotions) rather than design or evaluation decisions. These produce
-// time-bound state mutations, which is what isOperationalStateProgression keys
-// on. Deliberately excludes architecture/trade_off/design — those carry
-// genuine direction-setting forks that must remain detectable.
-var operationalTypes = map[string]bool{
-	"operations":  true,
-	"operational": true,
-	"deployment":  true,
-}
-
-// workItemScopedTypes are decision types whose conflict scope is a single work
-// item — a PR under review, a ticket being planned or investigated. Two such
-// decisions about DIFFERENT work items examine different code/scope and cannot
-// contradict each other, which is what isDisjointWorkItem keys on. Deliberately
-// excludes the direction-setting types (architecture / trade_off / design):
-// a cross-cutting design fork can legitimately span work items and must remain
-// detectable, the same reason operationalTypes excludes them.
-var workItemScopedTypes = map[string]bool{
-	"code_review":   true,
-	"review":        true,
-	"assessment":    true,
-	"investigation": true,
-	"analysis":      true,
-	"audit":         true,
-	"planning":      true,
-}
-
-// resourceScopedTypes are decision types whose conflict scope is a single
-// infrastructure resource — a connector being recovered, an org/customer
-// incident being investigated or deployed against. Two such decisions about
-// DIFFERENT resources act on different data planes and cannot contradict, which
-// is what isDisjointResource keys on (via extractResourceRefs). investigation
-// is deliberately a member of BOTH this set and workItemScopedTypes: an
-// investigation can be disjoint by ticket OR by connector, and either filter may
-// suppress it. Direction-setting types (architecture / trade_off / design) are
-// excluded for the same reason as the sibling type sets — a cross-cutting design
-// fork can legitimately span resources and must remain detectable.
-var resourceScopedTypes = map[string]bool{
-	"investigation":        true,
-	"deployment":           true,
-	"operational_recovery": true,
-	"operations":           true,
-	"operational":          true,
-	"incident":             true,
-}
-
-// isDirectionalWorkflowPair returns true if earlierType is a review/assessment
-// type and laterType is an implementation/fix type. Unlike isWorkflowPair in
-// validator.go (which checks both directions for LLM prompt hints), this check
-// is directional: the review must come first temporally.
-func isDirectionalWorkflowPair(earlierType, laterType string) bool {
-	return reviewTypes[strings.ToLower(earlierType)] && implementationTypes[strings.ToLower(laterType)]
-}
-
-// refinementKeywords are outcome substrings that indicate the decision
-// built on a prior decision by the same agent.
-var refinementKeywords = []string{
-	"implemented",
-	"fixed",
-	"resolved",
-	"completed",
-	"addressed",
-}
 
 func ticketContextNote(a, b []string) string {
 	if len(a) == 0 || len(b) == 0 {
